moved gcp_functions from scripts to src
